requirements.txt
Django==4.2
mysqlclient
transformers
torch
spacy
pyspellchecker